Move combine_distributions to univariate.py
